Trying to get the version checking working
There is a problem with the size of the version checked online... I have to add a 0 at the end of the version string or compare does not work because of some garbage. The conversion to char* seems to not work in that case, if a C++ guru have a better solution, try it.